+++ api/v3/StripeCustomer.php
@@ -5,6 +5,8 @@
  *
  */
 
+use CRM_Stripe_ExtensionUtil as E;
+
 /**
  * StripeCustomer.Get API specification
  *
@@ -80,6 +82,41 @@ function civicrm_api3_stripe_customer_get($params) {
 }
 
 /**
+ * StripeCustomer.Get API specification
+ *
+ * @param array $spec description of fields supported by this API call
+ * @return void
+ * @see http://wiki.civicrm.org/confluence/display/CRMDOC/API+Architecture+Standards
+ */
+function _civicrm_api3_stripe_customer_delete_spec(&$spec) {
+  $spec['id']['title'] = ts("Stripe Customer ID");
+  $spec['id']['type'] = CRM_Utils_Type::T_STRING;
+  $spec['contact_id']['title'] = ts("CiviCRM Contact ID");
+  $spec['contact_id']['type'] = CRM_Utils_Type::T_INT;
+  $spec['is_live']['title'] = ts("Is live processor");
+  $spec['is_live']['type'] = CRM_Utils_Type::T_BOOLEAN;
+  $spec['is_live']['api.required'] = TRUE;
+  $spec['processor_id']['title'] = ts("Payment Processor ID");
+  $spec['processor_id']['type'] = CRM_Utils_Type::T_INT;
+  $spec['processor_id']['api.required'] = TRUE;
+}
+
+/**
+ * StripeCustomer.delete API
+ *  This api will delete a stripe customer from CiviCRM
+ *
+ * @param array $params
+ * @see civicrm_api3_create_success
+ *
+ * @throws \Civi\Payment\Exception\PaymentProcessorException
+ * @return array
+ */
+function civicrm_api3_stripe_customer_delete($params) {
+  CRM_Stripe_Customer::delete($params);
+  return civicrm_api3_create_success([]);
+}
+
+/**
  * Stripe.Customer.Updatecontactids API
  *  This api will update the civicrm_stripe_customers table and add contact IDs for all known email addresses
  *
@@ -152,4 +189,86 @@ function civicrm_api3_stripe_customer_updatecontactids($params) {
   }
 
   return civicrm_api3_create_success($counts);
+}
+
+function _civicrm_api3_stripe_customer_updatestripemetadata_spec(&$spec) {
+  $spec['id']['title'] = E::ts("Stripe Customer ID");
+  $spec['id']['description'] = E::ts('If set only this customer will be updated, otherwise we try and update ALL customers');
+  $spec['id']['type'] = CRM_Utils_Type::T_STRING;
+  $spec['id']['api.required'] = FALSE;
+  $spec['dryrun']['api.required'] = TRUE;
+  $spec['dryrun']['type'] = CRM_Utils_Type::T_BOOLEAN;
+  $spec['is_live']['api.required'] = FALSE;
+  $spec['is_live']['type'] = CRM_Utils_Type::T_BOOLEAN;
+  $spec['processor_id']['api.required'] = FALSE;
+  $spec['processor_id']['type'] = CRM_Utils_Type::T_INT;
+}
+
+/**
+ * This allows us to update the metadata held by stripe about our CiviCRM payments
+ * Older versions of stripe extension did not set anything useful in stripe except email
+ * Now we set a description including the name + metadata holding contact id.
+ *
+ * @param $params
+ *
+ * @return array
+ * @throws \CiviCRM_API3_Exception
+ * @throws \Civi\Payment\Exception\PaymentProcessorException
+ */
+function civicrm_api3_stripe_customer_updatestripemetadata($params) {
+  if (!isset($params['dryrun'])) {
+    throw new CiviCRM_API3_Exception('Missing required parameter dryrun');
+  }
+  // Check params
+  if (empty($params['id'])) {
+    // We're doing an update on all stripe customers
+    if (!isset($params['is_live']) || !isset($params['processor_id'])) {
+      throw new CiviCRM_API3_Exception('Missing required parameters is_live and/or processor_id when using without a customer id');
+    }
+    $customerIds = CRM_Stripe_Customer::getAll($params['is_live'], $params['processor_id'], $params['options']);
+  }
+  else {
+    $customerIds = [$params['id']];
+  }
+  foreach ($customerIds as $customerId) {
+    $customerParams = CRM_Stripe_Customer::getParamsForCustomerId($customerId);
+    if (empty($customerParams['contact_id'])) {
+      throw new CiviCRM_API3_Exception('Could not find contact ID for stripe customer: ' . $customerId);
+    }
+
+    $paymentProcessor = \Civi\Payment\System::singleton()
+      ->getById($customerParams['processor_id']);
+    $paymentProcessor->setAPIParams();
+
+    // Get the stripe customer from stripe
+    try {
+      $stripeCustomer = \Stripe\Customer::retrieve($customerId);
+    } catch (Exception $e) {
+      $err = CRM_Core_Payment_Stripe::parseStripeException('retrieve_customer', $e, FALSE);
+      $errorMessage = CRM_Core_Payment_Stripe::handleErrorNotification($err, NULL);
+      throw new \Civi\Payment\Exception\PaymentProcessorException('Failed to retrieve Stripe Customer: ' . $errorMessage);
+    }
+
+    // Get the contact display name
+    $contactDisplayName = civicrm_api3('Contact', 'getvalue', [
+      'return' => 'display_name',
+      'id' => $customerParams['contact_id'],
+    ]);
+
+    // Currently we set the description and metadata
+    $stripeCustomerParams = [
+      'description' => $contactDisplayName . ' (CiviCRM)',
+      'metadata' => ['civicrm_contact_id' => $customerParams['contact_id']],
+    ];
+
+    // Update the stripe customer object at stripe
+    if (!$params['dryrun']) {
+      \Stripe\Customer::update($customerId, $stripeCustomerParams);
+      $results[] = $stripeCustomerParams;
+    }
+    else {
+      $results[] = $stripeCustomerParams;
+    }
+  }
+  return civicrm_api3_create_success($results, $params);
 }